Version 5.2.7
moved the location of the tool text to the upper right hand corner of the box. 
changed the tool text to be button specific by moving the <div>s to be children of the <nav> <Li> elements. 
put in <noscript> text to explain why the number of days will be "x" if JavaScript is unavailable